Config Enum Category. Additional, add config dto

+++ src/main/java/com/opticarlos/application/exceptions/ProductNotFoundException.java
@@ -2,7 +2,7 @@ package com.opticarlos.application.exceptions;
 
 public class ProductNotFoundException extends RuntimeException {
 
-    public ProductNotFoundException(Integer productId) {
+    public ProductNotFoundException(Long productId) {
         super("Product not found with ID: " + productId);
     }
